requirements.txt
asn1crypto==0.24.0
attrs==19.1.0
Automat==0.7.0
beautifulsoup4==4.8.0
cffi==1.12.3
constantly==15.1.0
cryptography==2.7
cssselect==1.0.3
Django==2.2.3
django-localflavor==2.2
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
lxml==4.3.4
parsel==1.5.1
psycopg2
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pyOpenSSL==19.0.0
python-stdnum==1.11
pytz==2019.1
queuelib==1.5.0
Scrapy==1.7.1
service-identity==18.1.0
six==1.12.0
soupsieve==1.9.2
sqlparse==0.3.0
Twisted==19.2.1
w3lib==1.20.0
zope.interface==4.6.0
